ms-pago y ms-factura

+++ ms-pago/src/main/java/com/example/mspago/repository/PagoRepository.java
@@ -1,9 +1,9 @@
 package com.example.mspago.repository;
+
 import com.example.mspago.entity.Pago;
-
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagoRepository extends JpaRepository<Pago, Long> {
+public interface PagoRepository extends JpaRepository<Pago, Integer> {
+
 
 }
